Display an information and tutorial dialog on first open.
This change adds a dialog window which, if no previous version of
btcgui has run before, will open before the main application window
opens and present the user with tips and steps to use the program.
After closing the dialog, the main application window will be openend
and run.

When the main application window opens, the current application
version will be written out to a file so it can be read on the next
application run.  Writing out the whole version also allows
implementing additional information dialogs in the future describing
changes relevant to users of any previous version.