Make sending transactions a little more intuitive.
This change fixes a issues regarding sending transaction.  First of
all, if the wallet is locked when trying to send a tx, an obscure
error message (about sending a JSON method to unlock the wallet) is
replaced with the with the dialog to unlock the wallet.  If the dialog
is cancelled out, attempting to send the transaction is not attempted.
However, if unlocking is successful, btcgui will try to requst the tx
creation and send a second time.

Second, if sending the tx succeeds, the widgets holding the addresses
and amounts are reset back to how they appear on first start (with a
single, not filled in, recipient).

While here, fixed an issue with wallet lock state notification not
using the new per-account notifications.