Disable widgets that don't make sense.
This disables the new wallet dialog in the settings menu, as new
wallet creation can no longer be canceled out of at first start if no
wallet exists.

The "Set Transaction Fee" option under the settings menu is also now
set to insensitive when btcwallet is disconnected, and sensitive when
connected.