Connect callback to correct signal.
When force closing the new wallet dialog, the callback must attach to
the "delete-event" signal, and not the "destroy" signal.  A
delete-event signal will be sent if the user requests the window to be
closed (closing it from the windwo manager), but destroy will signal
for anytime the window is destroyed, including after successfully
creating a wallet.  Connecting to the correct signal here keeps btcgui
open after entering a passphrase and creating the initial wallet.