Remove unneeded explicit requests.
Account balances and blockchain heights no longer need to be
explicitly requested.  Notifications updating each will be sent from
btcwallet when it is most appropiate (such as if btcgui is connected
to wallet when btcd connects to btcwallet, or if btcgui connects to a
btcwallet that is already connected to btcd).  If the result of the
notification cannot be determined due to the current connection
states, the calculation is not attempted.