Display transaction info.
This change adds transaction details to both the overview and
transaction panes.  All transactions are requested from btcwallet and
shown at startup, and any new transactions are added to the GUI via
btcwallet notifications.

Because btcwallet does not yet cleanly abort failed tx sends, this
code assumes that all added txs are never removed.  This will be
updated in the future to modify existing tx details as btcwallet makes
changes (for example, due to a txmined notification from btcd) or for
cleanly aborting a tx send.