Use smarter btcwallet cert path logic.
If ~/.btcgui/btcwallet.cert does not exist and the CA file has not
been explicitly set using the config file or command line flags, it's
possible that the cert can be found in ~/.btcwallet.  If connecting to
a localhost btcwallet and the previous statements are true, the
default CA file config option is updated for the certificate in them
btcwallet homedir.

If ~/.btcgui/btcwallet.cert does exist and the CA file has not been
set, it is used without checking for a cert in the btcwallet homedir.