Implement correct config/flag parsing for a GUI.
For a GUI, any command line flags should only be shown in an error
dialog if there was an actual error during the config or flag parsing.
go-flags treats help (-h) as an error, and entire help usage was being
printed to the pre-GUI error dialog.  This change fixes this behavior
to only show parsing errors and not help usage to the dialog.

Both help usage and error parsing remains printed to os.Stderr.