Rename connect option to rpcconnect.
The connect option is already used by btcd to force a connection to
other full node peers.  Wallet GUIs do not talk directly with these
peers, so the connect option is being renamed to something unique for
an RPC client connection.

This also matches the recent btcwallet change of renaming connect to
rpcconnect, for the same reasons as stated above.